requirements.txt
fastapi
python-chess
sqlalchemy
psycopg2-binary
requests
pytest
openai
